[#166] refactor: 홈페이지 추가 변경
