+++ src/app/_features/home/hooks/useNotice.ts
@@ -1,5 +1,5 @@
-import { getNotice } from '@/app/_services/getNotice';
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
+import { getNotice } from '../services/getNotice';
 
 export type NoticeType = 'notice' | 'notice-patch' | 'notice-event';
